:recycle: refactor: migrate csv utils to `src/shared`
