+++ api/lib/domain/services/certification-candidates-ods-service.js
@@ -2,9 +2,9 @@ import bluebird from 'bluebird';
 import _ from 'lodash';
 
 import { ComplementaryCertificationKeys } from '../../../src/certification/shared/domain/models/ComplementaryCertificationKeys.js';
+import * as readOdsUtils from '../../../src/shared/infrastructure/utils/ods/read-ods-utils.js';
 import * as mailCheckImplementation from '../../../src/shared/mail/infrastructure/services/mail-check.js';
 import { getTransformationStructsForPixCertifCandidatesImport } from '../../infrastructure/files/candidates-import/candidates-import-transformation-structures.js';
-import * as readOdsUtils from '../../infrastructure/utils/ods/read-ods-utils.js';
 import { CERTIFICATION_CANDIDATES_ERRORS } from '../constants/certification-candidates-errors.js';
 import { CertificationCandidatesError } from '../errors.js';
 import { CertificationCandidate } from '../models/CertificationCandidate.js';
